download osm data on the fly, with a cache
Data downloaded from osm are transformed in an in-between format
"OSMPoint" that has only the field we want. This format is stored in
the cache.

The cache consists of many little tiles ("atoms" in the code) so that
we can minimize the bounding box for new requests. This could probably
be further optimized (n^2).

BoundingBox struct, good for all 2D f64 bounding boxes.