setup interactive view to take the maximal size available
This kind of works but causes problems:
- the aspect ratio is broken
- this is extremly slow with long tracks (berlin)

To make the speed bearable for development i disable the link, that's
why i had to update test data. I also extend run.sh with a test that
exhibits how slow this is and make a flamegraph: most time is spend in
building the label placement graph in the hit
detection (Candidate::hit_other).